starting oop refactor
